Fixed refactor
